+++ packages/tools/atlas/src/bin/base/NUTS/run.sh
@@ -4,9 +4,10 @@ echo '
 ==== NUTS ====
 ' &&
 
-node -r esm src/bin/base/NUTS/01_getMetadata.js &&
-node -r esm src/bin/base/NUTS/01_getShapes.js &&
-node -r esm src/bin/base/NUTS/01-2_setUnifiedId.js &&
-node -r esm src/bin/base/NUTS/12_getCountriesTopojson.js &&
+node -r esm src/bin/base/NUTS/0-1_getBaseShapes.js &&
+node -r esm src/bin/base/NUTS/0-1_getMetadata.js &&
+node -r esm src/bin/base/NUTS/01-2_reconcileIds.js &&
+node -r esm src/bin/base/NUTS/12-3_augmentShapes.js &&
+node -r esm src/bin/base/NUTS/13-4_getCountriesTopojson.js &&
 
 echo