+++ packages/tools/atlas/CHANGELOG.md
@@ -1,8 +1,20 @@
 ## `@svizzle/atlas` v0.6.0 (next)
 
-- NUTS: add hierarchy and unified ids
+- NUTS:
+	- add resolution `10M`, update some texts in `3M` files
+	- reconcile recoded regions ids by introducing an 'unified' `atlasId`:
+	- provide static maps to facilitate in-app reconciliation:
+		- `unifiedNuts.json`: a list of regions with added:
+			- `id`: the atlas id
+			- `rootId`: the root region id (e.g. usually the country id)
+			- `pid`: the parent region id
+		- `hierarchy.json`: a hierarchy of regions based on the unified ids
+		- `yearlyNutsIdToId.json`: a map from `year/NUTS_ID` to `atlasId`
+		- `idToNutsIdByYear.json`: a map from `atlasId` to `NUTS_ID` by `year`
+		- `recoded.json`: a map of recodes as declared by NUTS (see `/atlas/data/base/NUTS/0/history/recoded` & `/atlas/data/base/NUTS/0/history/README.md`)
+	- augmented each region of the topojsons with a property `atlasId`
+- add the `getAtlasId` function
 - remove the dependency on `@svizzle/geo`
-- NUTS, add resolution `10M`, update some texts in `3M` files
 
 ## `@svizzle/atlas` v0.5.0